功能：编写 jsonrpc 服务端应用代码

+++ user/device/pwm_controller.h
@@ -16,8 +16,8 @@
 #define PCA9685_BASIC_DEFAULT_SUBADDRESS_2                0xE4                                              /**< 0xE4 */
 #define PCA9685_BASIC_DEFAULT_SUBADDRESS_3                0xE8                                              /**< 0xE8 */
 
-uint8_t pwm_controller_init     (pca9685_address_t addr, uint16_t hz);
-uint8_t pwm_controller_deinit   (void);
-uint8_t pwm_controller_write    (pca9685_channel_t channel, float delay_percent, float high_duty_cycle_percent);
+int pwm_controller_init     (pca9685_address_t addr, uint16_t hz);
+int pwm_controller_deinit   (void);
+int pwm_controller_write    (pca9685_channel_t channel, float delay_percent, float high_duty_cycle_percent);
 
 #endif
